Simplify `launchOrConnectToServer` to eliminate nested retries (#5872)
Previously we needed a `retryWithTimeout` around the reading of the port
because the initial process launch could take >10s, which would make the
outer `retryWithTimeout` not take effect. This is because we were
performing the coursier runner classpath resolution as part of that
launch, which can take arbitrarily long, overshooting the retry timeout.

This PR fixes it by moving the coursier runner classpath resolution out
of the retry loop, so the retry loop only contains the logic around
launching the process. This should help keep the time of each attempt to
a more reasonable amount, allowing the outer retries to work and
avoiding the need for nested retries around the individual steps of the
process launch workflow

---------

+++ testkit/src/mill/testkit/IntegrationTesterBase.scala
@@ -6,6 +6,7 @@ import mill.util.Retry
 trait IntegrationTesterBase {
   def workspaceSourcePath: os.Path
   def daemonMode: Boolean
+  def cleanupProcessIdFile: Boolean
 
   def propagateJavaHome: Boolean
 
@@ -75,7 +76,7 @@ trait IntegrationTesterBase {
   def removeProcessIdFile(): Unit = {
     if (!sys.env.contains("MILL_TEST_SHARED_OUTPUT_DIR")) {
       val outDir = os.Path(out, workspacePath)
-      if (os.exists(outDir)) {
+      if (os.exists(outDir) && cleanupProcessIdFile) {
         if (daemonMode) {
           val serverPath = outDir / millDaemon
           os.remove(serverPath / processId)
